[NCL-2871] Don't report missing permissions as err
Currently repour reports `git push` failures with exit code 1. This exit
code is interpreted by Maitai as a SYSTEM_ERROR. We want to change this
to a 'build error' instead by returning exit code 10 instead (See NCL-2079)

This commit modifies the exit_code of the git push exception (if any) to
10, and improves the description error by telling the user it does not
have the required permissions to push.