[NCL-2896] Dont shw stacktrace for failed callback
Repour logs when a callback fails to be completed is outputing the full
stacktrace. This is annoying since it doesn't provide that much
information, we just really need to know the reason of the failed
callback.

This commit removes the full stacktrace and instead just prints the
reason of the failed callback. In my opinion, this will help increase
the readability of the logs.

See the JIRA for the full stacktrace.