git push fails because tag already exists, is fine
Because we do not create branches anymore for /adjust, I had to change
the git push command to:

```
git push --tags
```

Sometimes this command will fail if the tag already exists on the remote
server (this situation happens because we do a shallow clone, and that
clone doesn't have all the tags in the remote server). This is fine
because that means that the tag that already exists will have the same
content as the tag created locally. So it's ok that the push failed.