[NCL-2882] Send live logs of repour operations
The live logs will only work when the operation is done in callback
mode. In callback mode, the callback id is sent to the requestor. The
requestor can then use that callback id to connect to a websocket
endpoint and receive live logs.

The websocket endpoint is:

```
ws://<server>:<port>/callback/{callback_id}
```

This is achieved by adding a logger handler that receives the logs, and
sends them to the websocket clients.

There is another logger handler used to capture the logs for specific
callback ids and save them into individual files. The aim is to be able
to forward to the requestor the relevant logs for that particular callback
id, beyond using websockets.

There is also a function that is activated. Its goal is to cleanup the
closed websockets that repour is sending logs if the websockets are
already closed. This helps prevent any memory leak that might happen if
we keep track of the closed websockets.