[NCL-2882] Send back websocket link to requestor
This commit requires the user to define an Environment variable named
'REPOUR_URL'. The value of that variable will be the external url of
repour, for e.g 'repour.somewhere.in.the.cloud'

The websocket link is only sent when callback mode is used.

The callback reply will look like:
```
{
    "callback": {
        "id": "fancy-id",
        "websocket": "ws://pnc-repour.instance.magic/callback/fancy-id"
    }
}
```