[NCL-3041] Update to handle scm user on its own
The Git url that we receive from the request might not have the username
in it anymore. This makes sense because Repour is configured to clone
repositories from a specific user, as repour is running with the private
keys of that specific user.

As such, the username of the git url we receive is _always_ for that
specific user. So it makes sense to just omit the username in the git
url we receive, and let repour add it implicitly instead. In that way,
if the specific user repour uses has to change, we only have to do the
change in repour, and not in the various clients of repour.