[NCL-2981] Add endpoint /cloneadjust
This endpoint allows a user to sync and optionally run PME on the
output.

If the ref is a branch, the branch will be pushed to the target repo. If
the ref is a tag, the tag will be pushed to the target repo. If the ref
is a SHA, a branch will be created and named 'branch-<ref>' and pushed
to the target repo. The latter is required to get a particular commit to
the target repository.

The adjust parameter, when enabled, will run PME, and the result is put
inside the response in key 'adjust_result'. The content of
'adjust_result' is exactly the same as the output for the '/adjust'
endpoint.