[Test] Fixes to test_adjust.py
The `adjust` function seems to have moved around, and this has probably
broken some tests. I also removed the branch references since we don't
explicitly create branches anymore.

Fixed a bug with the username being added to the url; if the username is
empty, it won't try to add the username (back to the old behaviour)