Wrong exit code used for tests
The previous exit code logic assumed that the `result.failures`
attribute returned an int, which indicated the number of failures.

This resulted in the comparison `result.failures != 0` to always be
true, whether there were failures or not, as the types of the 2 items we
are comparing are different.

`result.failures` actually returns a list of failed tests. As such, we
should instead use the sum of the length of the `result.failures` and
`result.errors` list as the exit code.

The exit code number now reflects the number of failures + errors in the
tests.