[NCL-2981] Adapt /adjust to have syncing feature
That sync feature is only enabled if the keys:

- originRepoUrl *and*
- sync (with value True)

are sent to the `/adjust` endpoint. Otherwise the sync won't happen and
`/adjust` will behave as before.

While I did a similar commit with adding the `/cloneadjust` endpoint, I
now realize that using the `/cloneadjust` endpoint on the Maitai process
is awkward because, based on the 'sync' being true or false, I'd have to
call `/cloneadjust` or `/adjust` endpoint respectively on the Maitai
side.

I believe it's better to add the feature to the `/adjust` endpoint so
that it requires minimum changes on the Maitai side, while keeping
`/adjust` backwards compatible.

A future commit will remove the `/cloneadjust` endpoint.