[NCL-3117] Unify syncing behaviour for /clone and /adjust
The previous behaviour of `/clone` was to create a branch if the ref
wasn't a branch. The original idea was to allow the user to manually
apply additional changes to the branch if required.

However, it turns out we never really needed that feature. This commit
removes this behaviour.

For both /clone and /adjust (with auto-sync enabled), if the ref is a:

- tag => tag gets synced with internal repo
- branch => branch gets synced with internal repo
- SHA => SHA gets synced with internal repo (via indirect creation of a
         tag with format 'repour-sync-<ref>' to make this possible)