[NCL-3159] Create tag that matches ver for /adjust
Repour currently creates a commit and tag after and adjustment is done
to a repository. The tag name is using the format 'repour-<commitid>'.

This commit attempts to change the tag name format to
'<version>'. The new tag format is more meaningful and should
help humans better understand what the tag is used for.

Since the new tag name format might hide who created the tag, the tag
annotation has a text that specifies that the tag was created by Repour.

The version information is extracted only when we use PME for the
adjustment phase. PME generates a result JSON file that contains the
version of the Java project after processing. That version is extracted
from the JSON file and is used to generate the new tag name.