[NCL-3503] endpoint adjusted to accept new param
The `/adjust` endpoint is adjust to now also accept the parameters:

- tempBuild [optional] (with as value a boolean)
- tempBuildTimestamp [optional] (with as value a string)

For `tempBuildTimestamp` to have any effect, `tempBuild` key must be
specified and must have a value of 'true'.

If `tempBuild` is set to true, PME will use a specific Indy group for
PME alignmnent.

The new parameter to define in the repour config is:

```
{
    ...
    "temp_build_indy_group": "<indy group to use>",
    ...
}
```

This commit allows further work to be done on the Maitai side. When the
specifics on how PME will deal with temp build and temp build timestamp
is known, a further commit will be done to use the new parameters.