[NCL-3555] Add the commitid to tag if tag already exists
During the adjust phase, we'll use as tag name the version returned
after a PME run.

However there might be a scenario where the tag name we want to create
is already present. This commit attempts to detect that scenario, and
use as tag name '<original-tag-name>-<commit id of PME change>'

Also with this commit, we log if tag already exists or not