[NCL2167] Add informational endpoint in Repour
This commit adds support for an informational endpoint in Repour when
accessing repour on '/'.

To do so, we deactivated the authentication on the '/' endpoint only so
that anyone could access it.

The informational endpoint shows the Repour version (as defined in
`repour/__init__.py`) and the Git SHA, if the `.git` folder is in the
root folder of the application.

Instructions on what to update during a release is also now provided.