[NCL-3601] Correctly process non top-level pom.xml
There are some git repositories where the root pom.xml is not located in
the root location of the repository, but rather in a folder.

```
<folder>/pom.xml
```

PME allows you to specify the `--file=<folder>/pom.xml` param to be
able to run PME from the root of the git repository.

However, when PME is finished, repour tries to read the file
`target/pom-manip-ext-result.json` from the root git repository, instead
of from `<folder>/target/pom-manip-ext-result.json`

So we can fix this issue by either:

- change directory to the 'folder' in the '--file' param, and remove
   the '--file' param in the list of PME parameters. Then run PME from
   the 'folder' to replicate what we do

- Not remove the '--file' param in hte list of PME parameters, and use
  the 'folder' to properly grab the pom-manip-ext-result.json.

While the second option is much simpler, this commit implements the
first option. This is because of PME 2.11 and 2.12 suffering from a bug
when using the '--file' param (See PRODTASKS-361). With that param,
the generated pom-manip-ext-result.json has null as values for all its
keys.

```
{
    "VersioningState": {
        "executionRootModified":{
            "groupId":null,
            "artifactId":null,
            "version":null
        }
    }
}
```

With the first option, we avoid this PME bug also.