Push existing tag name for /adjust
If the tag name after the adjust alignment already exists in the
external repo (with pre-sync enabled), we still cannot be sure that the
tag is present in the internal repository, since we only push the tag we
originally want to sync.

This commit makes sure that we also push into the internal repo the tag
which contains the alignment changes, just in case.

Thanks to @jochrist for reporting this bug!