[NCL-3984] Use checkout -f when changing branch
On a fresh clone, sometimes we still get 'diffs' in the working tree
because of weird behavious between CRLF and LF file ending conversion
between windows and linux, and the .gitattributes file.

This prevents us from checking out the branch/tag we want. So by using
the '-f' option in checkout, this allows us change branch/tag and
discard the local changes.

See: https://help.github.com/articles/dealing-with-line-endings/