[NCL-4047] Err msg for github private repo perm
When cloning a private github repositories, the user should add a
particular github user so that Repour has permissions to do it. The
error message will tell the user what to do when such a case happens.

The particular github user is specified in the environment variable
'PRIVATE_GITHUB_USER'