[NCL-4082] Endpoint to translate external scm to internal
Description of endpoint
=======================

Endpoint is: /git-external-to-internal
The endpoint is not authenticated. Therefore any user can access it.

Request
--------

POST /git-external-to-internal
```

{"external_url": "<url>"}
```

Responses
--------
- Success!

HTTP status: 200
```
{
    "external_url": "<url>",
    "internal_url": "<internal-url>",
    "log": "<logs>"
}
```

- Failure

HTTP status: 500
```
{
    "error_traceback": "<error-traceback>",
    "error_type": "<error-type>",
    "log": "<logs>"
}
```

Configuration Changes
=====================
We need to add key 'git_url_internal_template' to the configuration.
This is used as the first part of the internal url.

e.g 'git_url_internal_template': 'git+ssh://<private-git-server>/gerrit/'