[NCL-4120] Repour append commit id to tag on devel mode
During testing of repour, we use the `/adjust` endpoint with sync on between the
official internal git repository (which we consider as the external
repo) and a test internal git repository. There is a particular case
when the syncing is problematic:

- test internal git repository is aligned from upstream tag 1.0.0 and repour
  creates tag 1.0.0-company-1

- some time later the official git repository is aligned from upstream
  tag 1.0.0 in official repour instance and creates tag 1.0.0-company-1

Then the next time we sync the official internal git repo to the test
internal git repository, it fails because we have the same tag
(1.0.0-company-1) pointing to different commitids! (alignment will be
done slightly differently between the test repour and production repour
instance)

So instead this commit adds the 'devel' mode to repour that we'll
activate in test repour instances. It is activated by adding to the
repour top-level JSON config:

```
{
    ...
    "mode": "devel"
}
```

By default the mode is assumed to be 'prod'. What this does is that on
using `/adjust` in the test repour instances, instead of creating a tag
with format '<version>', it'll create a tag for format
'<version>-<commit id>'. This will help prevent tag commitid collisions
the next time we try to sync repositories.