[NC-4231] Fix bug in repour syncing of branches
When syncing branches on the `/adjust` endpoint, we:

1. git clone the upstream repository and checkout the ref
2. remove the remote 'origin' and set remote 'origin' to the internal repository
3. push and sync the 'ref' to sync to internal repository
4. Run PME
5. Create a commit, tag and push to internal repository

The issue is that on step 2, removing the remote 'origin' affects step 3.

Step 3 tries to identify if the ref to sync is a branch or a tag or a
sha. The reasoning is that if it's a branch or a tag, we push them as-is
to the internal repository. If it's a sha, we need to push it as a tag
to the internal repository.

On Step 3, we assume by default that the branch to check is
'remotes/origin/<ref>'. Unfortunately since we removed the remote
'origin' on Step 2, step 3 thinks the ref is not a branch
(remotes/origin/<ref> does not exist anymore when the remote is
removed). It instead incorrectly considers the ref as a sha.

The fix for this is that on Step 2, instead of removing the remote
'origin', we rename it to 'origin_remote'. Then on Step 3, we check if
the branch exists from 'remotes/origin_remote/<ref>' instead.

Another question that might be asked is why don't we just set the remote
of the internal repository to another name, instead of using 'origin'.
The reason is because this will affect other parts of the 'adjust' code
in subtle ways if different conditions are hit (like if the ref does not
exist at all)