[NCL-4974] Add monitoring to Repour
We currently monitor endpoints requests and error by using the
Prometheus client, exposing the metrics in the `/metrics` endpoint. We
can optionally export the data to Graphite if environment variable
`GRAPHITE_SERVER` and `GRAPHITE_KEY`` are defined. `GRAPHITE_KEY` is the
prefix for the data, and is usually set to the url of the server.

You can also override the default port for the Graphite server with
`GRAPHITE_PORT`

The metrics monitored are:

- time request of `/adjust`, `/clone`, `/cancel`,
  `/external-to-internal`, `/`, and sending result to callback urls

  * This covers latency and traffic

- errors:
  * validation json error
  * 400 response
  * 500 response
  * can't send response to callback

- CPU, memory, GC
  * Covers saturation