[NCL-5096] Add support for distributed 'cancel'
All the work is done in the `cancel` module.

The `start_cancel_loop` and the `check_if_other_repour_replicas_cancelled`
work hand in hand. Those stuff are done to be able to run repour with
multiple replicas.

When a request to cancel a task comes in, we first check if the task is
present in the repour server which got the request.

If yes -> cancel that task and report back to the client

If no ->
- Create an indicator file in a shared location to indicate that we want to
  cancel a task
- Verify every PERIOD_CANCEL_LOOP_SLEEP seconds, for up to 10 times, if the
  indicator file got deleted
- if yes -> the cancel was successful
- if no -> cancel was unsuccessful. Delete the indicator file and tell the
  caller the cancel operation was unsuccessful

- that indicator file is seen by other repour replicas (via the shared
  location) in the 'start_cancel_loop' - the other repour replicas check
  their event loop, and if they found and cancelled the task, they
  delete the indicator file

The 'start_cancel_loop' runs every PERIOD_CANCEL_LOOP_SLEEP seconds to check