[NCL-6165] Add brewPullActive to `adjust` request
The `brewPullActive`, and the `tempBuild` flags will now be passed to
the alignment tools to determine what Indy group to talk to. We already
have the 'tempBuild' flag, so I'm just adding the `brewPullActive` one
for now.

The new `brewPullActive` flag is used to limit the usage of brew pull
and only be used as required. This is done to speed up builds.

The logic on how to use it for alignment will be done on another commit.
This commit is necessary so that we can work on the other bits (Maitai,
new Maitai, PNC) request DTOs.