[NCL-6882] Cut off logs after a certain size
Repour sends back the entire logs of an alignment in the callback.
However, we cannot send logs greater than a certain size because this
will cause OutOfMemory exceptions to the other services.

This commit:
- checks the size of the repour logs to send back no matter if the
  alignment succeeded or failed
- if the size is greater than the maximum allowed (say foo bytes), then
  we only return the last foo bytes of the repour logs since it's more
  likely it'll contain the most important information
- mark the alignment as failed

The max size log can be configured in the `repour_config.json` as:
```
{
    ...
    "max_log_size_bytes": 12345
}
```

If this setting is not specified, then we don't activate this feature