[NCLSUP-626] PME alignment: pass the --f option to PME directly
Before this commit, Repour would remove the '-f' or '--file' option from
the extra PME parameters and run PME from the folder specified in the
'-f' option. However the PME '-f' option can be used to specify either a
folder or a file (can be named anything, doesn't have to finish with
pom.xml)

This logic unfortunately broke alignment for some projects where the
traditional file that contains the Maven setup is not named pom.xml.

This commit changes this by:
- truly providing the --file option values to PME rather than hiding it
- extracting any folder from the '--file' option to guess where the
  results file will be generated