[NCLSUP-1136] Sanitize project path for gitlab
Gitlab has a particular regex format for group and project name. We need
to modify the path we receive to adhere to that format.

Most of the paths we receive are from Github and Github probably has
different naming requirements compared to Gitlab.

If the start or end of the group/project ends with a non-alpha-numeric
character, we'll just replace it with an underscore.