[NCLSUP-1166] Add delay before running alignment
We want to add a configurable delay before running alignment due to a
bug in Rex which doesn't account for the time it takes for a build to be
entirely stored in PNC.

The delay will give a chance for PNC to store the results before
progressing. Once the Rex bug is fixed, we can remove that code.

This is set in the config.json by specifying:
```
    "adjust_delay_seconds": 100
```